WIP: entropy based password detection
This is harder than it looks. So one probably wants to use bigrams as
the distributions of characters (e.g. letters) in a random password
might not be that much off, while bigrams will be off by a lot. But
also considering digits and punctuation is hard.

The numbers in this commit came from training sets from two large
ansible repos of ours, but they are very biased. So the numbers are
quite off and shouldn't be used. One probably wants to use available
bigrams data for English.

Lastly, how to combine that data into one measure of "entropy"?
Shannon entropy is the wrong measure here. I researched a bit and it
looks like there is no well established measure that can be used here.
What we have is, for each char, the probability of that char occurring
given the previous char. So we can certainly find out of the string
contains mostly "rather expected" characters or mostly "rather
unexpected" characters.